Local instances database
When adding new instances, don't allow hrid to be NULL (which violates
a DB constraint).  Use the actual hrid if possible (MARC source record
with hrid in field 001) and fall back to the instance ID.  Then, fetch
all instances for which the ID (not hrid) was used and fix them.